+++ src/hal/indi/mod.rs
@@ -6,8 +6,10 @@ mod xml_reader;
 mod xml_helper;
 mod error;
 mod events;
+mod property;
 
 pub use num_format::*;
 pub use connection::*;
 pub use drivers::*;
 pub use events::*;
+pub use property::*;
